+++ package.json
@@ -23,24 +23,24 @@
 	},
 	"devDependencies": {
 		"@nabla/vite-plugin-eslint": "^1.4.0",
-		"@types/node": "^17.0.43",
+		"@types/node": "18.0.0",
 		"@typescript-eslint/eslint-plugin": "^5.28.0",
 		"@typescript-eslint/parser": "^5.28.0",
 		"@vitejs/plugin-vue": "^2.3.3",
 		"@vue/eslint-config-typescript": "^11.0.0",
 		"autoprefixer": "^10.4.7",
-		"eslint": "^8.17.0",
+		"eslint": "^8.18.0",
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-prettier": "^4.0.0",
 		"eslint-plugin-vue": "^9.1.1",
 		"husky": "^8.0.1",
 		"postcss": "^8.4.14",
-		"prettier": "^2.7.0",
+		"prettier": "^2.7.1",
 		"prettier-plugin-tailwindcss": "^0.1.11",
 		"tailwindcss": "^3.1.3",
-		"typescript": "^4.7.3",
+		"typescript": "^4.7.4",
 		"vite": "^2.9.12",
 		"vue-eslint-parser": "^9.0.2",
-		"vue-tsc": "^0.37.8"
+		"vue-tsc": "0.38.0"
 	}
 }
